+++ test/test.js
@@ -2172,12 +2172,12 @@ describe("Netmath", () => {
         })
 
         it("Weights are all between -0.5 and +0.5 when fanIn is 12", () => {
-            const result = NetMath.lecunUniform(10, {fanIn: 12})
+            const result = NetMath.lecunUniform(1000, {fanIn: 12})
             expect(result.every(w => w>=-0.5 && w<=0.5)).to.be.true
         })
 
         it("Inits some weights at values bigger |0.5| when fanIn is smaller (8)", () => {
-            const result = NetMath.lecunUniform(10, {fanIn: 8})
+            const result = NetMath.lecunUniform(1000, {fanIn: 8})
             expect(result.some(w => w<=-0.05 || w>=0.05)).to.be.true
         })
 
@@ -2216,6 +2216,41 @@ describe("Netmath", () => {
             const result = NetMath.xavierNormal(10, {fanIn: 5})
             expect(NetMath.lecunNormal).to.have.been.calledWith(10, {fanIn: 5})
             expect(result.length).to.equal(10)
+            NetMath.lecunNormal.restore()
+        })
+    })
+
+    describe("xavierUniform", () => {
+        it("Returns the same number of values as the size value given", () => {
+            const result = NetMath.xavierUniform(10, {fanIn: 10})
+            expect(result.length).to.equal(10)
+        })
+
+        it("Weights are all between -0.5 and +0.5 when fanIn is 10 and fanOut is 15", () => {
+            const result = NetMath.xavierUniform(1000, {fanIn: 10, fanOut: 15})
+            expect(result.every(w => w>=-0.5 && w<=0.5)).to.be.true
+        })
+
+        it("Inits some weights at values bigger |0.5| when fanIn+fanOut is smaller (5+5=10)", () => {
+            const result = NetMath.xavierUniform(1000, {fanIn: 5, fanOut: 5})
+            expect(result.some(w => w<=-0.05 || w>=0.05)).to.be.true
+        })
+
+        it("Creates weights that are more or less uniform", () => {
+            const result = NetMath.xavierUniform(1000, {fanIn: 5, fanOut: 5})
+
+            const decData = {}
+            result.forEach(w => decData[Math.abs(w*10).toString()[0]] = (decData[Math.abs(w*10).toString()[0]]|0) + 1)
+            const sorted = Object.keys(decData).map(k=>decData[k]).sort((a,b) => a<b)
+            expect(sorted[0] - sorted[sorted.length-1]).to.be.at.most(200)
+        })
+
+        it("Falls back to using lecunNormal if there is no fanOut available", () => {
+            sinon.spy(NetMath, "lecunUniform")
+            const result = NetMath.xavierUniform(10, {fanIn: 5})
+            expect(NetMath.lecunUniform).to.have.been.calledWith(10, {fanIn: 5})
+            expect(result.length).to.equal(10)
+            NetMath.lecunUniform.restore()
         })
     })
 })